third_party/blink/web_tests/external/wpt/css/cssom/CSSKeyframesRule-expected.txt

This is a testharness.js-based test.
[FAIL] name, cssRules, appendRule, findRule, deleteRule
  assert_equals: CSSKeyframesRule cssText attribute with CSS-wide keyword name expected "@keyframes\\"initial\\"{}" but got "@keyframesinitial{}"
Harness: the test ran to completion.
